Create Stochastic Gradient Solution
The goal of this project is to use machine learning to solve a problem that arises in Information Retrieval,
one known as the Learning to Rank (LeToR) problem. We formulate this as a problem of linear regression
where we map an input vector x to a real-valued scalar target y(x;w). 
Train a linear regression model on the LeToR dataset and synthetic dataset using stochastic gradient descent (SGD).